% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aoi.R
\name{aoi}
\alias{aoi}
\alias{set_aoi}
\alias{view_aoi}
\alias{get_aoi}
\title{Session-wide area of interest (AOI)}
\usage{
set_aoi(aoi)

view_aoi(color = "green")

get_aoi(type = "sf")
}
\arguments{
\item{aoi}{nothing, if an interactve \code{mapedit} viewer should open letting you draw an AOI polygon. Otherwise, sfc_POLYGON or SpatialPolygons or matrix, representing a single multi-point (at least three points) polygon of your area-of-interest (AOI). If it is a matrix, it has to have two columns (longitude and latitude) and at least three rows (each row representing one corner coordinate). If its projection is not \code{+proj=longlat +datum=WGS84 +no_defs}, it is reprojected to the latter.}

\item{color}{chracter, polygon filling color.}

\item{type}{character, AOI object type, either "matrix", "sf" or "sp".}
}
\description{
Functions that set, view and get a session-wide area of interest (AOI) that can be used by all \code{getSpatialData} functions.
}
\details{
\code{set_aoi} defines a session AOI that is used for querying data within the running session (if no other AOI is provided with a query function call). If called without argument, an interactive \code{mapedit} viewer is opened letting you draw an AOI polygon. Otherwise, the function supports \code{sf}, \code{sp} or matrix objects as \code{aoi} input  (see argument \code{aoi}).

\code{view_aoi} displays the defined session AOI on an interactive \code{mapview}/\code{leaflet} map.

\code{get_aoi} returns the defined session AOI.
}
\examples{

## set aoi (example aoi)
data("aoi_data")
set_aoi(aoi_data[[1]])

\dontrun{
## draw aoi interactively
set_aoi() # just call without an argument

## view aoi
view_aoi()

## return aoi
aoi <- get_aoi(type = "sf")
}

}
\seealso{
\link{get_records}
}
\author{
Jakob Schwalb-Willmann
}
