% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_hot_cloudcov.R
\name{calc_hot_cloudcov}
\alias{calc_hot_cloudcov}
\title{Calculates the haze-optimal transformation cloud cover based on the red and blue band}
\usage{
calc_hot_cloudcov(record, preview, aoi = NULL, identifier = NULL,
  maxDeviation = 20, sceneCloudCoverCol, cloudPrbThreshold = 40,
  slopeDefault = 1.4, interceptDefault = -10, dir_out = NULL,
  num_records = 1, verbose = TRUE)
}
\arguments{
\item{record}{data.frame, single line representing one record from a records data.frame.}

\item{preview}{raster, subject of cloud cover calculation. Either two layers: layer 1 = red, layer 2 = blue. Or three layers: layer 1 = red, layer 2 = something, layer 3 = blue.}

\item{aoi}{sp or sf, the aoi.}

\item{identifier}{numeric, column number where a unique identifier of the scenes is located, sensor-specific.}

\item{maxDeviation}{numeric, the maximum allowed deviation of calculated scene cloud cover from the provided scene cloud cover. Use 100 if you do not like to consider the cloud cover % given by the data distributor. Default is \code{maxDeviation = 20}.}

\item{sceneCloudCoverCol}{character, the clear name of the column in the record data.frame where the cloud cover estimation of the data dissiminator is found.}

\item{cloudPrbThreshold}{numeric, the threshold of the HOT cloud probability layer (0-100 %) below which pixels are considered as clear sky. Default is \code{cloudPrbThreshold = 40}. It will be dynamically adjusted according to the input in \code{maxDeviation}.}

\item{slopeDefault}{numeric, value taken as slope ONLY if least-alternate deviation regression fails.  Default is 1.4, proven to work well for common land surfaces.f default values. In this case cloud cover will be set to 9999 % for the given record.}

\item{dir_out}{character, optional. Full path to target directory where to save the cloud masks. If \code{NULL}, cloud masks are not saved.}

\item{verbose}{logical, if \code{TRUE}, details on the function's progress will be visibile on the console. Default is TRUE.}

\item{intercept}{Default, value taken as intercept ONLY if least-alternate deviation regression fails. Default is -10, proven to work well for common land surfaces.}
}
\value{
A data.frame, one line as the input with one additional column holding the estimated cloud cover within the aoi.
}
\description{
\code{calc_hot_cloudcov} estimates the cloud cover of a satellite image raster using the haze-optimal transformation (HOT) within an aoi
The algorithm was implemented in this function for cloud cover estimation of DN values from preview images.
}
\details{
The estimation of the cloud cover is done on the red and blue information of the input RGB. Haze-optimal transformation (HOT) procedure is applied based on 
Zhu & Helmer (2018), https://data.fs.usda.gov/research/pubs/iitf/ja_iitf_2018_Zhu.pdf. Orignally, the algorithm was introduced by Zhang et al. (2002)
"An image transform to characterize and compensate for spatial variations in thin cloud contamination of Landsat images", Remote Sensing of Environment 82, 2-3.
HOT seperates clear-sky pixels first from a threshold, calculates a least alternate deviation (LAD) regression from these pixels and exposes cloud pixels by the deviation of all pixels from this clear-sky line.
}
\note{
Mainly thought for internal use and only tested for preview images.
}
\author{
Henrik Fisser
}
