% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcMODIS_aoi_cloudcov.R
\name{calcMODIS_aoi_cloudcov}
\alias{calcMODIS_aoi_cloudcov}
\title{Calculate the cloud cover of MODIS data within an aoi}
\usage{
calcMODIS_aoi_cloudcov(records, aoi = NULL, maxDeviation = 20,
  cloudPrbThreshold = 40, slopeDefault = 1.4, interceptDefault = -10,
  dir_out = NULL, username = NULL, password = NULL, verbose = TRUE)
}
\value{
A data.frame as the records input with one additional column holding the estimated cloud cover within the aoi
}
\description{
\code{calcMODIS_aoi_cloudcov} estimates the cloud cover of Landsat data based on preview images using the haze-optimal transformation (HOT)
}
\details{
The estimation of the cloud cover is done on the red and blue information of the preview images provided by the respective data dissiminator.
}
\seealso{
\link{getMODIS_query} \link{getMODIS_preview} \link{getMODIS_data}
}
\author{
Henrik Fisser
}
