% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_availability.R
\name{check_availability}
\alias{check_availability}
\title{Check download availability of records}
\usage{
check_availability(records, verbose = TRUE)
}
\arguments{
\item{records}{records data frame, containing one or multiple records (each represented by one row), as returned by \link{get_records}}

\item{verbose}{logical, whether to display details on the function's progress or output on the console.}
}
\value{
A data frame of records (as defined with argument \code{records}), extended by a column \code{download_available} (logical).
}
\description{
\code{check_availability} checks for each record whether it is available for direct download (can be downloaded instantly) or not (and thus must be ordered before download).
}
\note{
To use this function, you must be logged in at the services required for your request. See the examples and \link{login} for details.
}
\author{
Jakob Schwalb-Willmann
}
