% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crop_file.R
\name{crop_file}
\alias{crop_file}
\title{Crop to spatial extent}
\usage{
crop_file(file, ext, verbose = TRUE, ...)
}
\arguments{
\item{file}{Path to raster file.}

\item{ext}{Extent object or such from which an extent can be extracted.}

\item{verbose}{logical, if \code{TRUE}, details on the function's progress will be visibile on the console. Default is TRUE.}

\item{...}{additional arguments passed to \code{writeRaster}, e.g. \code{filename}.}
}
\value{
Raster* object.
}
\description{
\code{crop_file} crops a raster file to a spatial extent using GDAL. It is useful when working with large-scale, memory-intensive datasets.
}
\details{
GDAL must be installed. If no GDAL installation is found, \code{crop_file} is masking \code{raster::crop}.
}
\seealso{
\link{getSentinel_data}
}
\author{
Jakob Schwalb-Willmann
}
