% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cropFAST.R
\name{cropFAST}
\alias{cropFAST}
\title{Crop file to a spatial extent (beta)}
\usage{
cropFAST(file, ext, filename, verbose = TRUE, ...)
}
\arguments{
\item{file}{Path to raster file.}

\item{ext}{Extent object or such from which an extent can be extracted.}

\item{filename}{Filename of output, if it should be written to disk.}

\item{...}{additional arguments passed to \code{writeRaster}. Ignored, if \code{filename} ends with ".vrt".}
}
\value{
Raster* object.
}
\description{
\code{cropFAST} crops a raster file to a spatial extent using GDAL. It is useful when working with large-scale, memory-intensive datasets.
}
\details{
GDAL must be installed. If no GDAL installation is found, \code{cropFAST} is masking \code{raster::crop}.
The input file on disk will not be changed.
}
\examples{
\dontrun{
library(raster)

## load a file
path_tiff <- "/path/to/file.tiff"
r_tiff <- stack(path_tiff)

## define an AOI
data("aoi_data")
aoi <- aoi_data[[2]] # example AOI

## if necessary, reproject AOI to file CRS
aoi <- spTransform(aoi, crs(r))

## crop (file on disk will not be changed)
r_crop <- cropFAST(path_tiff, ext = aoi)

## crop to a new file
r_crop <- cropFAST(path_tiff, ext = aoi, filename = "/path/to/file_cropped.tiff")
}

}
\seealso{
\link{getSentinel_data}
}
\author{
Jakob Schwalb-Willmann
}
