% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{.make_mosaic}
\alias{.make_mosaic}
\title{create mosaic}
\usage{
.make_mosaic(x, save_path, mode = "mask", srcnodata = NULL,
  datatype = NULL)
}
\arguments{
\item{x}{list of paths to raster files.}

\item{save_path}{character, full path where to save the mosaic (has to end with '.tif').}

\item{mode}{character, optional. If mode == "rgb" no masking of the raster to only 1 values is done. 
If mode == "mask" the raster will be returned with 1 and NA. Default is mode == "mask".}

\item{srcnodata}{character nodata value in x. Default is for FLT4S.}
}
\value{
\code{mos} raster mosaic
}
\description{
The rasters from \code{x} will be mosaicked in a stupid way: everything is mosaicked that is in this list.
}
\keyword{internal}
