% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{.preview_mask_edges}
\alias{.preview_mask_edges}
\title{mask the edges of Landsat preview raster}
\usage{
.preview_mask_edges(preview)
}
\arguments{
\item{preview}{raster.}
}
\value{
\code{preview_masked} masked preview
}
\description{
mask the edges of Landsat preview raster
}
