% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cloudcov_supported.R
\name{get_cloudcov_supported}
\alias{get_cloudcov_supported}
\title{Returns product names supported by \code{calc_cloudcov}}
\usage{
get_cloudcov_supported()
}
\value{
character vector of product names supported by \link{calc_cloudcov}.
}
\description{
\code{get_cloudcov_supported} provides all product names supported by \link{calc_cloudcov}.
Note that among the Sentinel-3 sensors only OLCI products are supported.
}
\author{
Henrik Fisser, 2020
}
