% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\alias{getSentinel_data}
\alias{getLandsat_data}
\alias{getMODIS_data}
\title{Download datasets}
\usage{
get_data(
  records,
  dir_out = NULL,
  md5_check = TRUE,
  force = FALSE,
  as_sf = TRUE,
  ...,
  verbose = TRUE
)

getSentinel_data(
  records,
  dir_out = NULL,
  md5_check = TRUE,
  force = FALSE,
  as_sf = TRUE,
  ...,
  verbose = TRUE
)

getLandsat_data(
  records,
  dir_out = NULL,
  md5_check = TRUE,
  force = FALSE,
  as_sf = TRUE,
  ...,
  verbose = TRUE
)

getMODIS_data(
  records,
  dir_out = NULL,
  md5_check = TRUE,
  force = FALSE,
  as_sf = TRUE,
  ...,
  verbose = TRUE
)
}
\arguments{
\item{records}{records data frame, containing one or multiple records (each represented by one row), as returned by \link{get_records}}

\item{dir_out}{character, a directory, to which the downloaded file(s) should be saved. By default, these are saved to the archive directory defined with \code{set_archive}.}

\item{md5_check}{logical, whether to check md5 checksums (if available) or not.}

\item{force}{logical, whether to force download of files already existing or not.}

\item{as_sf}{logical, whether records should be returned as \code{sf} \code{data.frame} or a simple \code{data.frame}. In both cases, spatial geometries are stored in column \code{footprint}.}

\item{...}{additional, sensor-specific arguments:
\itemize{
   \item \code{hub}, character, Copernicus Hub selection for Sentinel queries. Either
\itemize{
   \item "auto" (default) to automatically select a suitable Copernicus hub depending on the selected products
   \item "dhus" to look for operational Open Hub records only,
   \item "s5p" to look for Sentinel-5P precursor pre-operational records only,
   \item "GNSS" to look for GNSS RINEX records only,
   \item or a valid API URL.
}
}}

\item{verbose}{logical, whether to display details on the function's progress or output on the console.}
}
\value{
A data frame of records (as defined with argument \code{records}), extended by additional columns.
}
\description{
\code{get_data} downloads the full datasets per records. File paths are added to the records data frame.
}
\details{
Sentinel data are downloaded from the *ESA Copernicus Open Access Hubs*.

Landsat data are downloaded from *USGS-EROS ESPA* (on-demand higher-level data) and *Amazon Web Srvices* (Landsat-8 Level 1 data).

MODIS data are downloaded from the Level-1 and Atmosphere Archive & Distribution System (LAADS) of NASA's Distributed Active Archive Center (DAAC) at the Goddard Space Flight Center in Greenbelt, Maryland (\url{https://ladsweb.modaps.eosdis.nasa.gov/}).
}
\note{
To use this function, you must be logged in at the services required for your request. See the examples and \link{login} for details.
}
\author{
Jakob Schwalb-Willmann
}
