% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_names.R
\name{get_names}
\alias{get_names}
\alias{getSentinel_names}
\alias{getLandsat_names}
\alias{getMODIS_names}
\title{Get product names}
\usage{
get_names(grouped = FALSE)

getSentinel_names()

getLandsat_names(username = NULL, password = NULL)

getMODIS_names(username = NULL, password = NULL)
}
\arguments{
\item{username}{character, user name to login at the respective service.}

\item{password}{character, password to login at the respective service.}
}
\value{
A character vector
}
\description{
\code{get_names} obtains names of available products from included services. The returned names can be used with the \link{get_records} functions for querying a specific product.
}
\examples{
\dontrun{
library(getSpatialData)

# login
login_USGS(username = "your_username")
login_CopHub(username = "your_username")

# get all available products
products <- get_names()
products$Sentinel
products$Landsat
products$MODIS

# or get products by service
products <- getSentinel_names()
products <- getLandsat_names()
products <- getMODIS_names()
}
}
\seealso{
getSentinel_records getLandsat_records getMODIS_records
}
\author{
Jakob Schwalb-Willmann
}
