% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_products.R
\name{get_products}
\alias{get_products}
\alias{getSentinel_products}
\alias{getLandsat_products}
\alias{getMODIS_products}
\title{Get product names}
\usage{
get_products(grouped = FALSE)

getSentinel_products()

getLandsat_products(username = NULL, password = NULL)

getMODIS_products(username = NULL, password = NULL)
}
\arguments{
\item{grouped}{logical, whether to return a list of names grouped by platform/sensor (\code{TRUE}) or not.}

\item{username}{character, user name to login at the respective service.}

\item{password}{character, password to login at the respective service.}
}
\value{
A character vector
}
\description{
\code{get_products} obtains the names of all products available through \code{getSpatialData}. The returned names can be used with the \link{get_records} functions for querying a specific product.
}
\note{
To use this function, you must be logged in at the services required for your request. See the examples and \link{login} for details.
}
\examples{
\dontrun{
library(getSpatialData)

# login
login_USGS(username = "your_username")
login_CopHub(username = "your_username")

# get all available products
products <- get_products()
products$Sentinel
products$Landsat
products$MODIS

# or get products by service
products <- getSentinel_products()
products <- getLandsat_products()
products <- getMODIS_products()
}
}
\seealso{
getSentinel_records getLandsat_records getMODIS_records
}
\author{
Jakob Schwalb-Willmann
}
