% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_records.R
\name{get_records_drivers}
\alias{get_records_drivers}
\title{Returns drivers for writing records}
\usage{
get_records_drivers()
}
\value{
list (named) of characters drivers (names) and their file extensions (values).
}
\description{
\code{get_records_drivers} provides the driver names that can be used in \link{write_records}.
}
\author{
Henrik Fisser, 2020
}
