% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_select_supported.R
\name{get_select_supported}
\alias{get_select_supported}
\title{Returns product names supported by \code{select_*} functionalities}
\usage{
get_select_supported()
}
\value{
character vector of product names supported by \code{select_*}.
}
\description{
\code{get_select_supported} provides all product names supported by:
\itemize{
\item \link{select_unitemporal}
\item \link{select_bitemporal}
\item \link{select_timeseries}
}
Other products than the returned cannot be processed by \code{select_*}.
Note that among the Sentinel-3 sensors only OLCI products are supported.
}
\author{
Henrik Fisser, 2020
}
