% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLandsat_data.R
\name{getLandsat_data}
\alias{getLandsat_data}
\title{Download Landsat data}
\usage{
getLandsat_data(products, level = "sr", source = "auto", dir_out = NULL,
  espa_order = NULL, force = FALSE, username = NULL, password = NULL,
  verbose = TRUE)
}
\arguments{
\item{products}{data.frame, one or multiple prodcuts (each represented by one row), as it is returned by \link{getLandsat_query}.}

\item{level}{character, the requested product level. Defaul is "sr" for surface reflectance. Available levels can be obtained from the "levels_available" field returned for each product by \link{getLandsat_query}.}

\item{source}{character, either:
\itemize{
   \item "auto" for automatic selection of data source depending on \code{level}
   \item "ESPA" to download on-demand products from USGS-EROS ESPA
   \item "AWS" to download from Amazon Webservices (Landsat-8 with \code{level="l1"} only)
}}

\item{dir_out}{character, full path to download target directory. Optional. If not set, \code{getSentinel_data} uses the directory to the \code{getSpatialData} archive folder. Use \link{set_archive} to once define a getSpatialData  archive folder.}

\item{espa_order}{character, optional. A vector of a single or multiple ESPA order IDs. Use this argument, if you want to download items being part of an order that already had been placed by this function or yourself earlier, e.g. in case you arboted the function while it was waiting for the order to be completed. The ESPA order ID is displayed when the order is placed and you recieve it via E-Mail from USGS-EROS. If defined, \code{products} is allowed to be undefined.}

\item{force}{logical. If \code{TRUE}, download is forced even if file already exisits in the download directory. Default is \code{FALSE}.}

\item{username}{character, a valid USGS user name. Default is NULL. Leave it undefined, if you want to use use \link{login_USGS} to define the login credentials once for all \code{get*} functions that connect to USGS services during the session. Register on \url{https://ers.cr.usgs.gov/register/}.}

\item{password}{character, the password to the specified user account. If \code{NULL}, the password will be taken from \link{login_USGS} inputs or, if \code{login_USGS} is not in use, asked interactively.}

\item{verbose}{logical, if \code{TRUE}, details on the function's progress will be visibile on the console. Default is TRUE.}
}
\value{
Character vector of files that had been downloaded.
}
\description{
\code{getLandsat_data} downloads Landsat datasets queried using \link{getLandsat_query} from different sources.
}
\note{
ESPA is used as source if higher-level products are requested by the user (see \code{level}). Since ESPA is an on-demand service, \code{getLandsat_data} places an order and then waits for the requested items to be available, before they are downloaded. Therefore, the runtime of the function is depending on how fast an order is being processed by the ESPA server. The ESPA processing time depends on the size of the order and can take up to 48 hours in highly demanding cases! The function status is indicated by the console messages that it is prompting during execution.
}
\seealso{
\link{getLandsat_names} \link{getLandsat_query} \link{getLandsat_preview}
}
\author{
Jakob Schwalb-Willmann
}
