% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLandsat_preview.R
\name{getLandsat_preview}
\alias{getLandsat_preview}
\title{Preview a Landsat product}
\usage{
getLandsat_preview(product, on_map = TRUE, show_aoi = TRUE,
  username = NULL, password = NULL, verbose = TRUE)
}
\arguments{
\item{product}{data.frame, single row data.frame collected from the return of \link{getLandsat_query}, representing the selected product and all its attributes.}

\item{on_map}{logical, if \code{TRUE}, the preview is displaed corner-georeferenced on a map. If \code{FALSE}, a simple RGB plot is displayed. Default is \code{TRUE}.}

\item{show_aoi}{logical, if \code{TRUE}, the session AOI defined with \link{set_aoi} is drawn to the map viewer. Ignored, if \code{on_map = FALSE} or if no AOI has been defined with \code{set_aoi}. Default is \code{TRUE}.}

\item{username}{character, a valid user name to the Copernicus Open Access Hub. Default is NULL. Leave it undefined, if you want to use use \link{login_CopHub} to define the login credentials once for all \code{getSentinel*} functions during the session. Register on \url{https://scihub.copernicus.eu/}. Not needed for connections to the pre-operational hub.}

\item{password}{character, the password to the specified user account. If \code{NULL}, the password will be taken from \link{login_CopHub} inputs or, if \code{login_CopHub} is not in use, asked interactively. Not needed for connections to the pre-operational hub.}

\item{verbose}{logical, if \code{TRUE}, details on the function's progress will be visibile on the console. Default is TRUE.}
}
\value{
None. A plot/view display is generated.
}
\description{
\code{getLandsat_preview} previews single products as RGB plot which had been queried using \link{getLandsat_query}. The function is useful to apply visual checks to products before downloading them.
}
\examples{

}
\seealso{
\link{getLandsat_names} \link{getLandsat_query} \link{getLandsat_data}
}
\author{
Jakob Schwalb-Willmann
}
