% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLandsat_query.R
\name{getLandsat_query}
\alias{getLandsat_query}
\title{Query Landsat data}
\usage{
getLandsat_query(time_range, name, aoi = NULL, username = NULL,
  password = NULL)
}
\arguments{
\item{time_range}{character, containing two elements: the query's starting date and stopping date, formatted "YYYY-MM-DD", e.g. "2017-05-15"}

\item{name}{character, identifies the name of the dataset to be queried. Use \link{getLandsat_names} to revcieve a vector with all available Landsat datasets from Earth Explorer.}

\item{aoi}{sfc_POLYGON or SpatialPolygons or matrix, representing a single multi-point (at least three points) polygon of your area-of-interest (AOI). If it is a matrix, it has to have two columns (longitude and latitude) and at least three rows (each row representing one corner coordinate). If its projection is not \code{+proj=longlat +datum=WGS84 +no_defs}, it is reprojected to the latter. Use \link{set_aoi} instead to once define an AOI globally for all queries within the running session. If \code{aoi} is undefined, the AOI that has been set using \link{set_aoi} is used.}

\item{username}{character, a valid USGS user name. Default is NULL. Leave it undefined, if you want to use use \link{login_USGS} to define the login credentials once for all \code{get*} functions that connect to USGS services during the session. Register on \url{https://ers.cr.usgs.gov/register/}.}

\item{password}{character, the password to the specified user account. If \code{NULL}, the password will be taken from \link{login_USGS} inputs or, if \code{login_USGS} is not in use, asked interactively.}
}
\value{
A data frame; each row represents one dataset. The data frame can be further filtered by its columnwise attributes. The selected rows can be handed over to the other getLandsat functions for previeqing or downloading.
}
\description{
\code{getLandsat_query} queries the USGS Earth Explorer for Landsat data by some basic input search parameters. The function returns a data frame that can be further filtered.
}
\seealso{
\link{getLandsat_names} \link{getLandsat_preview} \link{getLandsat_data}
}
\author{
Jakob Schwalb-Willmann
}
