% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMODIS_preview.R
\name{getMODIS_preview}
\alias{getMODIS_preview}
\title{Preview a MOSIS image}
\usage{
getMODIS_preview(record, on_map = TRUE, show_aoi = TRUE,
  verbose = TRUE)
}
\arguments{
\item{record}{data.frame, single row data.frame collected from the return of \link{getMODIS_query}, representing the selected record and all its attributes.}

\item{on_map}{logical, if \code{TRUE}, the preview is displaed corner-georeferenced on a map. If \code{FALSE}, a simple RGB plot is displayed. Default is \code{TRUE}.}

\item{show_aoi}{logical, if \code{TRUE}, the session AOI defined with \link{set_aoi} is drawn to the map viewer. Ignored, if \code{on_map = FALSE} or if no AOI has been defined with \code{set_aoi}. Default is \code{TRUE}.}

\item{verbose}{logical, whether to display details on the function's progress or output on the console.}
}
\value{
None. A plot/view display is generated.
}
\description{
\code{getMODIS_preview} previews a single MODIS record as image on a map or as RGB plot. The function is useful to apply visual checks to records before downloading them.
}
\seealso{
\link{getMODIS_names} \link{getMODIS_query} \link{getMODIS_data}
}
\author{
Jakob Schwalb-Willmann
}
