% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSentinel_restore.R
\name{getSentinel_restore}
\alias{getSentinel_restore}
\title{Restore Sentinel datasets from Copernicus LTA}
\usage{
getSentinel_restore(record, username = NULL, password = NULL,
  hub = "auto", verbose = TRUE)
}
\arguments{
\item{record}{data.frame, single row data.frame collected from the return of \link{getSentinel_records}, representing the selected record and all its attributes.}

\item{username}{to be removed.}

\item{password}{to be removed.}

\item{verbose}{logical, whether to display details on the function's progress or output on the console.}
}
\value{
TRUE on success.
}
\description{
\code{getSentinel_restore} requests to restore Setninel datasets that have been archived by ESA to the Copernicus Long-Term Archive (LTA) (see argument \code{check_avail} of \link{getSentinel_records}).
}
\seealso{
\link{getSentinel_records}
}
\author{
Jakob Schwalb-Willmann
}
