% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gSD_login.R
\name{gSD_login}
\alias{gSD_login}
\alias{login_CopHub}
\alias{login_USGS}
\title{Session-wide login}
\usage{
login_CopHub(username, password = NULL)

login_USGS(username, password = NULL)
}
\arguments{
\item{username}{character, username to the corresponding service. For details on registration, see details.}

\item{password}{character, password to the corresponding service.}
}
\value{
None.
}
\description{
Login to services supported by \code{getSpatialData} once for the running session.
}
\details{
\code{login_CopHub} defines login credentials for the Copernicus Open Access Hub (register on \url{https://scihub.copernicus.eu/})

\code{login_USGS} defines login credentials to the USGS EROS Registration System (ERS). Register on \url{https://ers.cr.usgs.gov/register/}.
}
\note{
Login credentials will be saved and made available for all \code{getSpatialData} functions during the whole session. They will be erased when quitting the session. Alternatively, login credentials can be set individually with each \code{get*} function call.
}
\examples{
library(getSpatialData)

\dontrun{
## Define user credentials for the Copernicus Open Access Hub
login_CopHub(username = "my_user_name", password = "my_password")

## Define USGS user credentials
login_USGS(username = "my_user_name", password = "my_password")
}
}
\seealso{
getSentinel_query getLandsat_query
}
\author{
Jakob Schwalb-Willmann
}
