% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gSD_settings.R
\name{gSD_settings}
\alias{gSD_settings}
\alias{login_CopHub}
\alias{login_USGS}
\alias{set_archive}
\alias{set_aoi}
\alias{view_aoi}
\title{getSpatialData settings}
\usage{
login_CopHub(username, password = NULL)

login_USGS(username, password = NULL)

set_archive(dir_archive)

set_aoi(aoi)

view_aoi(color = "green")
}
\arguments{
\item{username}{character, username to the corresponding service. For details on registration, see details.}

\item{password}{character, password to the corresponding service.}

\item{dir_archive}{character, directory to the \code{getSpatialData} archive folder.}

\item{aoi}{nothing, if an interactve \code{mapedit} viewer should open letting you draw an AOI polygon. Otherwise, sfc_POLYGON or SpatialPolygons or matrix, representing a single multi-point (at least three points) polygon of your area-of-interest (AOI). If it is a matrix, it has to have two columns (longitude and latitude) and at least three rows (each row representing one corner coordinate). If its projection is not \code{+proj=longlat +datum=WGS84 +no_defs}, it is reprojected to the latter.}

\item{color}{chracter, polygon filling color.}
}
\value{
None.
}
\description{
Defines global settings that are used by all \code{getSpatialData} functions
}
\details{
\code{login_CopHub} defines the login credentials for the Copernicus Open Access Hub (register on \url{https://scihub.copernicus.eu/}), \code{login_USGS} defines USGS login credentials Register on \url{https://ers.cr.usgs.gov/register/}. These will be saved and made available for all \code{getSpatialData} functions during the whole session and will be erased when quitting the session. Alternatively, login credentials can be set individually with each \code{get*} function call.

\code{set_archive} globally defines the directory on your machine (or an external device) where getSpatialData should build up its donwload data archive. Since getSpatialData handles big amounts of data, it is recommended to once define a location where enough free storage is available and then afterwards to not change the archive location. You need to define the archives location for each session after loading getSpatialData. It will then be remembered for the duration of the session. Apart from the archive location, you can manually define a download path when calling the *_data functions. If you do not define a path there, getSpatialData will direct the download to the defined archive. The archive is structred by sensors.

\code{set_aoi} globally defines an AOI that is used for querying data within the running session (if no other AOI is provided with a query function call). If called without argument, an interactive \code{mapedit} viewer is opened letting you draw an AOI polygon. Otherwise, the function supports \code{sf}, \code{sp} or matrix objects as \code{aoi} input  (see argument \code{aoi}).

\code{view_aoi} displays the defined AOI on an interactive \code{mapview}/\code{leaflet} map.
}
\examples{
## Define user credentials for the Copernicus Open Access Hub
login_CopHub(username = "my_user_name", password = "my_password")

## Define USGS user credentials
login_USGS(username = "my_user_name", password = "my_password")

## set archive directory
dir.arc <- tempdir()
set_archiv(dir.arc)

## set aoi (example aoi)
aoi <- data("aoi")[[1]]
set_aoi(aoi)

\dontrun{
## draw aoi interactively
set_aoi() # just call without an argument

## view aoi
view_aoi()
}

}
\seealso{
getSentinel_query
}
\author{
Jakob Schwalb-Willmann
}
