% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_product_checker.R
\name{is.landsat5}
\alias{is.landsat5}
\title{Returns TRUE for records that are of product 'LANDSAT_TM_C1' (Landsat-5)}
\usage{
is.landsat5(records)
}
\description{
\code{is.landsat5} checks which records are Landsat-5 records.
}
\author{
Henrik Fisser, 2020
}
