% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_product_checker.R
\name{is.landsatMSS}
\alias{is.landsatMSS}
\title{Returns TRUE for records that are of product group 'MODIS'
Returns TRUE for records that are of product 'LANDSAT_MSS_C1'}
\usage{
is.landsatMSS(records)
}
\description{
\code{is.landsatMSS} checks which records are Landsat MSS records.
}
\author{
Henrik Fisser, 2020
}
