% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_product_checker.R
\name{is.modis}
\alias{is.modis}
\title{Returns TRUE for records that are of product group 'MODIS'}
\usage{
is.modis(records)
}
\description{
\code{is.modis} checks which records are MODIS records.
}
\author{
Henrik Fisser, 2020
}
