% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_product_checker.R
\name{is.modis_aqua}
\alias{is.modis_aqua}
\title{Returns TRUE for records that are of product group 'MODIS' and sensor Aqua}
\usage{
is.modis_aqua(records)
}
\arguments{
\item{records}{data.frame, one or multiple records (each represented by one row), as it is 
returned by \link{get_records}.}
}
\value{
logical vector, same length as number of rows in \code{records}.
}
\description{
\code{is.modis_aqua} checks which records are MODIS Aqua records ('MYD')
}
\author{
Henrik Fisser, 2020
}
