% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_product_checker.R
\name{is.product}
\alias{is.product}
\title{Returns TRUE for records that are of the specified product}
\usage{
is.product(records, product)
}
\arguments{
\item{product}{character specifies the product to be checked on.}
}
\description{
\code{is.product} checks which records are of product \code{product}.
}
\details{
Check \link{get_names} for available product names.
}
\author{
Henrik Fisser, 2020
}
