% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_product_checker.R
\name{is.product_}
\alias{is.product_}
\title{Returns TRUE for records that are of the specified product}
\usage{
is.product_(records, product)
}
\arguments{
\item{records}{sf data.frame}

\item{product}{character specifies the product to be checked on.}
}
\value{
logical vector
}
\description{
\code{is.product_} checks which records are of product \code{product}.
}
\details{
Check \link{get_names} for available product names.
}
\author{
Henrik Fisser, 2020
}
