% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_product_checker.R
\name{is.product_group_}
\alias{is.product_group_}
\title{Returns TRUE for records that are of the specified product group}
\usage{
is.product_group_(records, product_group)
}
\arguments{
\item{records}{sf data.frame}

\item{product_group}{character specifies the product group to be checked on.}
}
\value{
logical vector
}
\description{
\code{is.product_group_} checks which records are of product group \code{product_group}
}
\author{
Henrik Fisser, 2020
}
