% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_product_checker.R
\name{is.sentinel}
\alias{is.sentinel}
\title{Returns TRUE for records that are of product group 'Sentinel'}
\usage{
is.sentinel(records)
}
\description{
\code{is.sentinel} checks which records are Sentinel records.
}
\author{
Henrik Fisser, 2020
}
