% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_product_checker.R
\name{is.sentinel1}
\alias{is.sentinel1}
\title{Returns TRUE for records that are of product 'Sentinel-1'}
\usage{
is.sentinel1(records)
}
\description{
\code{is.sentinel1} checks which records are Sentinel-1 records.
}
\author{
Henrik Fisser, 2020
}
