% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_product_checker.R
\name{is.sentinel1_iw_ocn}
\alias{is.sentinel1_iw_ocn}
\title{Returns TRUE for records that are of product 'Sentinel-1' IW OCN}
\usage{
is.sentinel1_iw_ocn(records)
}
\arguments{
\item{records}{data.frame, one or multiple records (each represented by one row), as it is 
returned by \link{get_records}.}
}
\value{
logical vector, same length as number of rows in \code{records}.
}
\description{
\code{is.sentinel1_iw_ocn} checks which records are Sentinel-1 Interferometric Wideswath (IW)
Ocean (OCN) records.
}
\author{
Henrik Fisser, 2020
}
