% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_product_checker.R
\name{is.sentinel2}
\alias{is.sentinel2}
\title{Returns TRUE for records that are of product 'Sentinel-2'}
\usage{
is.sentinel2(records)
}
\description{
\code{is.sentinel2} checks which records are Sentinel-2 records.
}
\author{
Henrik Fisser, 2020
}
