% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_product_checker.R
\name{is.sentinel2_S2A}
\alias{is.sentinel2_S2A}
\title{Returns TRUE for records that are of product 'Sentinel-2' and sensor S2A}
\usage{
is.sentinel2_S2A(records)
}
\arguments{
\item{records}{data.frame, one or multiple records (each represented by one row), as it is 
returned by \link{get_records}.}
}
\value{
logical vector, same length as number of rows in \code{records}.
}
\description{
\code{is.sentinel2_S2A} checks which records are Sentinel-2A records.
}
\author{
Henrik Fisser, 2020
}
