% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_product_checker.R
\name{is.sentinel3}
\alias{is.sentinel3}
\title{Returns TRUE for records that are of product 'Sentinel-3'}
\usage{
is.sentinel3(records)
}
\description{
\code{is.sentinel3} checks which records are Sentinel-3 records.
}
\author{
Henrik Fisser, 2020
}
