% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_product_checker.R
\name{is.sentinel3_olci}
\alias{is.sentinel3_olci}
\title{Returns TRUE for records that are of product 'Sentinel-3' OLCI}
\usage{
is.sentinel3_olci(records)
}
\description{
\code{is.sentinel3_olci} checks which records are Sentinel-3 OLCI records.
}
\author{
Henrik Fisser, 2020
}
