% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_product_checker.R
\name{is.sentinel3_S3C}
\alias{is.sentinel3_S3C}
\title{Returns TRUE for records that are of product 'Sentinel-3' and sensor S3C}
\usage{
is.sentinel3_S3C(records)
}
\arguments{
\item{records}{data.frame, one or multiple records (each represented by one row), as it is 
returned by \link{get_records}.}
}
\value{
logical vector, same length as number of rows in \code{records}.
}
\description{
\code{is.sentinel3_S3C} checks which records are Sentinel-3C records.
}
\author{
Henrik Fisser, 2020
}
