% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_product_checker.R
\name{is.sentinel3_slstr}
\alias{is.sentinel3_slstr}
\title{Returns TRUE for records that are of product 'Sentinel-3' SLSTR}
\usage{
is.sentinel3_slstr(records)
}
\description{
\code{is.sentinel3_slstr} checks which records are Sentinel-3 SLSTR records.
}
\author{
Henrik Fisser, 2020
}
