% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_product_checker.R
\name{is.sentinel5p}
\alias{is.sentinel5p}
\title{Returns TRUE for records that are of product 'Sentinel-5p'}
\usage{
is.sentinel5p(records)
}
\arguments{
\item{records}{data.frame, one or multiple records (each represented by one row), as it is 
returned by \link{get_records}.}
}
\value{
logical vector, same length as number of rows in \code{records}.
}
\description{
\code{is.sentinel5} checks which records are Sentinel-5 records.
}
\author{
Henrik Fisser, 2020
}
