% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_product_checker.R
\name{is_landsat5}
\alias{is_landsat5}
\title{Returns TRUE for records that are of product 'LANDSAT_TM_C1' (Landsat-5)}
\usage{
is_landsat5(records)
}
\arguments{
\item{records}{data.frame, one or multiple records (each represented by one row), as it is 
returned by \link{get_records}.}
}
\value{
logical vector, same length as number of rows in \code{records}.
}
\description{
\code{is_landsat5} checks which records are Landsat-5 records.
}
\author{
Henrik Fisser, 2020
}
