% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_product_checker.R
\name{is_landsat7}
\alias{is_landsat7}
\title{Returns TRUE for records that are of product 'LANDSAT_ETM_C1' (Landsat-7)}
\usage{
is_landsat7(records)
}
\arguments{
\item{records}{data.frame, one or multiple records (each represented by one row), as it is 
returned by \link{get_records}.}
}
\value{
logical vector, same length as number of rows in \code{records}.
}
\description{
\code{is_landsat7} checks which records are Landsat-7 records.
}
\author{
Henrik Fisser, 2020
}
