% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_product_checker.R
\name{is_modis}
\alias{is_modis}
\title{Returns TRUE for records that are of product group 'MODIS'}
\usage{
is_modis(records)
}
\arguments{
\item{records}{data.frame, one or multiple records (each represented by one row), as it is 
returned by \link{get_records}.}
}
\value{
logical vector, same length as number of rows in \code{records}.
}
\description{
\code{is_modis} checks which records are MODIS records.
}
\author{
Henrik Fisser, 2020
}
