% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_product_checker.R
\name{is_sentinel}
\alias{is_sentinel}
\title{Returns TRUE for records that are of product group 'Sentinel'}
\usage{
is_sentinel(records)
}
\arguments{
\item{records}{data.frame, one or multiple records (each represented by one row), as it is 
returned by \link{get_records}.}
}
\value{
logical vector, same length as number of rows in \code{records}.
}
\description{
\code{is_sentinel} checks which records are Sentinel records.
}
\author{
Henrik Fisser, 2020
}
