% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_product_checker.R
\name{is_sentinel2}
\alias{is_sentinel2}
\title{Returns TRUE for records that are of product 'Sentinel-2'}
\usage{
is_sentinel2(records)
}
\arguments{
\item{records}{data.frame, one or multiple records (each represented by one row), as it is 
returned by \link{get_records}.}
}
\value{
logical vector, same length as number of rows in \code{records}.
}
\description{
\code{is_sentinel2} checks which records are Sentinel-2 records.
}
\author{
Henrik Fisser, 2020
}
