% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login.R
\name{login}
\alias{login}
\alias{login_CopHub}
\alias{login_USGS}
\alias{services_avail}
\title{Session-wide login}
\usage{
login_CopHub(username = NULL, password = NULL, n_retry = 3,
  verbose = TRUE)

login_USGS(username = NULL, password = NULL, n_retry = 3,
  verbose = TRUE)

services_avail(value = F, verbose = T)
}
\arguments{
\item{username}{character, user name to login at the respective service.}

\item{password}{character, password to login at the respective service.}

\item{n_retry}{numeric, number of attempts to login, if login fails}

\item{verbose}{logical, whether to display details on the function's progress or output on the console.}

\item{value}{logical, whether to return a data frame containing service status or not. Default is \code{FALSE}.}
}
\description{
Functions to login to services supported by \code{getSpatialData} once for the running session and to recieve their current status.
}
\details{
\code{login_CopHub} defines login credentials for the Copernicus Open Access Hub (register on \url{https://scihub.copernicus.eu/})

\code{login_USGS} defines login credentials to the USGS EROS Registration System (ERS). Register on \url{https://ers.cr.usgs.gov/register/}.

\code{services_avail} displays the status of all online services used by \code{getSpatialData}. Services that are operating as usual are labeled "available". Returns a \code{data.frame} containing the service status, if argument \code{value} is set to \code{TRUE}.
}
\note{
Login credentials will be saved and made available for all \code{getSpatialData} functions during the whole session. They will be erased when quitting the session. Alternatively, login credentials can be set individually with each \code{get*} function call.
}
\examples{
library(getSpatialData)

\dontrun{
## Define user credentials for the Copernicus Open Access Hub
login_CopHub(username = "my_user_name", password = "my_password")

## Define USGS user credentials
login_USGS(username = "my_user_name", password = "my_password")

## get status of all services
services_avail()
}
}
\seealso{
\link{get_records}
}
\author{
Jakob Schwalb-Willmann
}
