% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_data.R
\name{order_data}
\alias{order_data}
\title{Order datasets}
\usage{
order_data(records, wait_for_order = FALSE, ..., verbose = TRUE)
}
\arguments{
\item{records}{records data frame, containing one or multiple records (each represented by one row), as returned by \link{get_records}}

\item{wait_for_order}{logical, whether to wait until all datasets have been successfully ordered and are ready for download (default) or not. If \code{FALSE}, orders are only placed.}

\item{...}{additional, sensor-specific arguments:
\itemize{
   \item \code{hub}, character, Copernicus Hub selection for Sentinel queries. Either
\itemize{
   \item "auto" (default) to automatically select a suitable Copernicus hub depending on the selected products
   \item "dhus" to look for operational Open Hub records only,
   \item "s5p" to look for Sentinel-5P precursor pre-operational records only,
   \item "GNSS" to look for GNSS RINEX records only,
   \item or a valid API URL.
}
}}

\item{verbose}{logical, whether to display details on the function's progress or output on the console.}
}
\value{
A data frame of records (as defined with argument \code{records}), extended by additional columns.
}
\description{
\code{order_data} oders datasets that are not available for immediate download (on-demand) but need to be ordered or restored before download. Use \link{check_availability} to see which datasets need to be ordered before download.
}
\note{
To use this function, you must be logged in at the services required for your request. See the examples and \link{login} for details.
}
\author{
Jakob Schwalb-Willmann
}
