% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepSentinel.R
\name{prepSentinel}
\alias{prepSentinel}
\title{Prepare Sentinel data}
\usage{
prepSentinel(dir_zip, dir_out = NULL, format = "vrt",
  select.tiles = "all", overwrite = F, verbose = T)
}
\arguments{
\item{dir_zip}{character, path to zip file.}

\item{dir_out}{character, full path to target directory. Optional. If not set, \code{getSentinel_data} uses the directory to the \code{getSpatialData} archive folder. Use \link{set_archive} to once define a getSpatialData  archive folder.}

\item{format}{character, output format of the raster datasets (indicated by the file extension, e.g. "tiff" for format GTiff). Default is "vrt".}

\item{select.tiles}{character, selecting one or multuple specific tiles to prepare. Older datasets archived by ESA contain multiple tiles. If \code{select.tiles} is set to one or multiple specific tiles contained within the selected datasets (e.g. "T32TNT"), only these tiles will be extracted and prepared. By default, all tiles are extracted and prepared. Ignored for datasets that contain only a single tile.}

\item{overwrite}{logical, whether to overwrite existing files or not.}

\item{verbose}{logical, if \code{TRUE}, details on the function's progress will be visibile on the console. Default is TRUE.}
}
\value{
Character vector of paths to the prepared files.
}
\description{
\code{prepSentinel} makes downloaded Sentinel datasets ready-to-use by automatically extracting, sorting and converting the relevant contents of the datasets to a user-defined format.
}
\seealso{
\link{getSentinel_data}
}
\author{
Jakob Schwalb-Willmann
}
