% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_previews.R
\name{read_previews}
\alias{read_previews}
\title{Read records previews}
\usage{
read_previews(records)
}
\arguments{
\item{records}{records data frame, containing one or multiple records (each represented by one row), as returned by \link{get_records}}
}
\value{
A list of \code{raster stack} objects
}
\description{
\code{read_previews} reads georeferences preview images downloaded using \code{get_previews} and returns them as a list of \code{raster stack} objects.
}
\author{
Jakob Schwalb-Willmann
}
