% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_records.R
\name{read_records}
\alias{read_records}
\title{read records from file}
\usage{
read_records(file, as_sf = TRUE, verbose = TRUE)
}
\arguments{
\item{file}{character file path from where to read the records file.}

\item{as_sf}{logical specifies if the records shall be returned as sf object.
Otherwise a data.frame will be returned. Default is TRUE.}

\item{verbose}{logical, if \code{TRUE}, details on the function's progress will be visibile 
on the console. Default is TRUE.}
}
\value{
sf object or data.frame depending on \code{as_sf}.
}
\description{
\code{read_records} reads records from a csv file
that has been previously written through \link{write_records}.
}
\author{
Henrik Fisser, 2020
}
