% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_python.R
\name{set_python}
\alias{set_python}
\title{Set Python path}
\usage{
set_python(py_path)
}
\arguments{
\item{py_path}{character, path to preferred python installation}
}
\value{
None.
}
\description{
Sets up the connection to a python installation manually.
}
\author{
Jakob Schwalb-Willmann
}
