% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_records.R
\name{view_records}
\alias{view_records}
\title{View footprints of records}
\usage{
view_records(records, show_aoi = TRUE, aoi_colour = "deepskyblue",
  value = FALSE, verbose = TRUE)
}
\arguments{
\item{records}{records data frame, containing one or multiple records (each represented by one row), as returned by \link{get_records}}

\item{show_aoi}{logical, whether the session AOI defined with \link{set_aoi} should be drawn on the map or plot(s) (\code{TRUE}) or not.}

\item{aoi_colour}{character, colour of the AOI. Ignored, if \code{show_aoi = FALSE}.}

\item{value}{logical, whether to return the displayed map/plot object (\code{TRUE}) or not.}

\item{verbose}{logical, whether to display details on the function's progress or output on the console.}
}
\value{
None or a \code{mapview}/\code{plot} object
}
\description{
\code{view_records} displays the footsprint of each record on a map.
}
\author{
Jakob Schwalb-Willmann
}
