% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_records.R
\name{write_records}
\alias{write_records}
\title{Writes records to a csv file}
\usage{
write_records(records, dir_out = NULL, file = NULL)
}
\arguments{
\item{dir_out}{directory where to write the records csv. A file name will be automatically
generated from date and time in case \code{dir_out} is set through \link{set_archive} or provided. 
Default is NULL.}

\item{file}{character complete file path where to write the records csv. Default is NULL.
If a \code{dir_out} is not NULL because it is set through \link{set_archive} or provided, 
\code{file} will be ignored.}
}
\value{
character file path to the saved records csv.
}
\description{
\code{write_records} write records to a csv file
that has been previously written through \link{write_records}.
}
\author{
Henrik Fisser, 2020
}
