% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\alias{getSentinel_data}
\alias{getLandsat_data}
\alias{getMODIS_data}
\title{Download datasets}
\usage{
get_data(records, dir_out = NULL, md5_check = TRUE, force = FALSE,
  ..., verbose = TRUE)

getSentinel_data(records, dir_out = NULL, md5_check = TRUE,
  force = FALSE, ..., verbose = TRUE)

getLandsat_data(records, dir_out = NULL, md5_check = TRUE,
  force = FALSE, ..., verbose = TRUE)

getMODIS_data(records, dir_out = NULL, md5_check = TRUE,
  force = FALSE, ..., verbose = TRUE)
}
\arguments{
\item{records}{records data frame, containing one or multiple records (each represented by one row), as returned by \link{get_records}}

\item{dir_out}{character, a directory, to which the downloaded file(s) should be saved. By default, these are saved to the archive directory defined with \code{set_archive}.}

\item{md5_check}{logical, whether to check md5 checksums (if available) or not.}

\item{force}{logical, whether to force download of files already existing or not.}

\item{...}{additional, sensor-specific arguments:
\itemize{
   \item \code{gnss}, logical, whether to query for Sentinel GNSS RINEX records instead of remote sensing instrument records. If \code{TRUE}, only records of the dual-frequency GPS recievers mounted on Sentinel-1, -2, and -3 are returned and \code{aoi} settings are ignored. If \code{FALSE} (default), remote sensing instrument records, queried including \code{aoi} settings, are returned (see section \code{Sentinel}).
   \item \code{hub}, character, Copernicus Hub selection for Sentinel queries. Either
\itemize{
   \item "auto" (default) to automatically select a suitable Copernicus hub depending on the selected products
   \item "dhus" to look for operational Open Hub records only,
   \item "s3" to look for Sentinel-3 pre-operational records only,
   \item "s5p" to look for Sentinel-5P precursor pre-operational records only,
   \item "GNSS" to look for GNSS RINEX records only,
   \item or a valid API URL.
}
}}

\item{verbose}{logical, whether to display details on the function's progress or output on the console.}
}
\value{
A data frame of records (as defined with argument \code{records}), extended by additional columns.
}
\description{
\code{get_data} downloads the full datasets per records. File paths are added to the records data frame.
}
\note{
To use this function, you must be logged in at the services required for your request. See the examples and \link{login} for details.
}
\author{
Jakob Schwalb-Willmann
}
