% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_names.R
\name{get_names}
\alias{get_names}
\alias{getSentinel_names}
\alias{getLandsat_names}
\alias{getMODIS_names}
\title{Get product names}
\usage{
get_names()

getSentinel_names()

getLandsat_names(username = NULL, password = NULL)

getMODIS_names(username = NULL, password = NULL)
}
\arguments{
\item{username}{character, a valid user name to the USGS EROS Registration System (ERS). If \code{NULL} (default), the session-wide login credentials are used (see \link{login_USGS} for details on registration).}

\item{password}{character, the password to the specified user account. If \code{NULL} (default) and no seesion-wide password is defined, it is asked interactively ((see \link{login_USGS} for details on registration).}
}
\value{
A character vector
}
\description{
\code{get_names} obtains names of available products from included services. The returned names can be used with the \link{get_query} functions for querying a specific product.
}
\examples{
\dontrun{
library(getSpatialData)

# login
login_USGS(username = "your_username")
login_CopHub(username = "your_username")

# get all available products
products <- get_names()
products$Sentinel
products$Landsat
products$MODIS

# or get products by service
products <- getSentinel_names()
products <- getLandsat_names()
products <- getMODIS_names()
}
}
\seealso{
getSentinel_query getLandsat_query getMODIS_query
}
\author{
Jakob Schwalb-Willmann
}
