% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLandsat_data.R
\name{getLandsat_data}
\alias{getLandsat_data}
\title{Download Landsat data}
\usage{
getLandsat_data(records, level = "sr", source = "auto",
  dir_out = NULL, wait_for_espa = NULL, force = FALSE,
  username = NULL, password = NULL, n.retry = 3, verbose = TRUE)
}
\arguments{
\item{records}{data.frame, one or multiple records (each represented by one row), as it is returned by \link{getLandsat_query}.}

\item{level}{character, one or multiple product levels to be requested. Defaul is "sr" for surface reflectance. Available levels can be obtained from the "levels_available" field returned for each product by \link{getLandsat_query}.}

\item{source}{character, either:
\itemize{
   \item "auto" for automatic selection of data source depending on \code{level}
   \item "ESPA" to download on-demand products from USGS-EROS ESPA
   \item "AWS" to download from Amazon Webservices (Landsat-8 with \code{level="l1"} only)
}}

\item{dir_out}{character, full path to download target directory. Optional. If not set, \code{getLandsat_data} uses the directory to the \code{getSpatialData} archive folder. Use \link{set_archive} to once define a getSpatialData  archive folder.}

\item{force}{logical. If \code{TRUE}, download is forced even if file already exisits in the download directory. Default is \code{FALSE}.}

\item{n.retry}{numeric, maximum number of download (re-)attempts. If downloads of datasets fail (e.g. MD5 checksums do not match), these downloads will be reattampted.}

\item{espa_order}{character, optional. A vector of a single or multiple ESPA order IDs. Use this argument, if you want to download items being part of an order that already had been placed by this function or yourself earlier, e.g. in case you arboted the function while it was waiting for the order to be completed. The ESPA order ID is displayed when the order is placed and you recieve it via E-Mail from USGS-EROS. If defined, \code{records} is allowed to be undefined.}
}
\value{
Character vector of paths to the downloaded files.
}
\description{
\code{getLandsat_data} downloads Landsat data queried using \link{getLandsat_query} from different sources.
}
\note{
ESPA is used as source if higher-level products are requested by the user (see \code{level}). Since ESPA is an on-demand service, \code{getLandsat_data} places an order and then waits for the requested items to be available, before they are downloaded. Therefore, the runtime of the function is depending on how fast an order is being processed by the ESPA server. The ESPA processing time depends on the size of the order and can take up to 48 hours in highly demanding cases! The function status is indicated by the console messages that it is prompting during execution.
}
\examples{

## Load packages
library(getSpatialData)
library(sf)

## set aoi and time range for the query
set_aoi(aoi_data[[1]])
time_range <-  c("2017-08-01", "2017-08-30")

## Login to USGS ERS
\dontrun{
login_USGS("username")

## set archive directory
set_archive("/path/to/archive/")

## get available products and select one
product_names <- getLandsat_names()

## query for records for your AOI, time range and product
query <- getLandsat_query(time_range = time_range, name = product_names[7])

## preview a record
getLandsat_preview(query[5,])

#print available levels for a record
query[5,]$levels_available

## download record 5 with level "l1" (will direct to AWS automaticaly)
files <- getLandsat_data(records = query[5,], level = "l1", source = "auto")

## download record 5 with level "sr" (will be processed on demand by ESPA)
files <- getLandsat_data(records = query[5,], level = "sr", source = "auto")
# this can take very long, since the function will wait,
# until the processing by ESPA is done

## you can abort the function while it is waiting for ESPA and resume later:
files <- getLandsat_data(espa_order = "espa-XYZA@host.com-YOUR-ORDER-ID")
# the order IDs are displayed and send by mail, use them to resume the task
}


}
\seealso{
\link{getLandsat_names} \link{getLandsat_query} \link{getLandsat_preview}
}
\author{
Jakob Schwalb-Willmann
}
