% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLandsat_names.R
\name{getLandsat_names}
\alias{getLandsat_names}
\title{Get Landsat dataset names from the USGS Earth Explorer}
\usage{
getLandsat_names(username = NULL, password = NULL)
}
\arguments{
\item{username}{character, a valid USGS user name. Default is NULL. Leave it undefined, if you want to use \link{login_USGS} to define the login credentials once for all \code{get*} functions that connect to USGS services during the session. Register on \url{https://ers.cr.usgs.gov/register/}.}

\item{password}{character, the password to the specified user account. If \code{NULL}, the password will be taken from \link{login_USGS} inputs or, if \code{login_USGS} is not in use, asked interactively.}
}
\value{
A character vector
}
\description{
\code{getLandsat_names} obtains names of available Landsat datasets from the USGS Earth Explorer. They can be used with the \link{getLandsat_query} function for querying a Landsat data.
}
\seealso{
\link{getLandsat_query}
}
\author{
Jakob Schwalb-Willmann
}
