% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLandsat_query.R
\name{getLandsat_query}
\alias{getLandsat_query}
\title{Query Landsat data}
\usage{
getLandsat_query(time_range, name = "all", aoi = NULL, username = NULL,
  password = NULL, verbose = TRUE)
}
\arguments{
\item{time_range}{character, containing two elements: the query's starting date and stopping date, formatted "YYYY-MM-DD", e.g. "2017-05-15"}

\item{name}{character, optional. Identifies the name of the dataset to be queried. If set to "all" (default), every available Landsat dataset is searched for results and included in the output. Use \link{getLandsat_names} to revcieve a vector with all available Landsat datasets from Earth Explorer, if you want to select a specific one.}

\item{aoi}{sfc_POLYGON or SpatialPolygons or matrix, representing a single multi-point (at least three points) polygon of your area-of-interest (AOI). If it is a matrix, it has to have two columns (longitude and latitude) and at least three rows (each row representing one corner coordinate). If its projection is not \code{+proj=longlat +datum=WGS84 +no_defs}, it is reprojected to the latter. Use \link{set_aoi} instead to once define an AOI globally for all queries within the running session. If \code{aoi} is undefined, the AOI that has been set using \link{set_aoi} is used.}

\item{username}{character, a valid user name to the Copernicus Open Access Hub. Default is NULL. Leave it undefined, if you want to use use \link{login_CopHub} to define the login credentials once for all \code{getSentinel*} functions during the session. Register on \url{https://scihub.copernicus.eu/}. Not needed for connections to the pre-operational hub.}

\item{password}{character, the password to the specified user account. If \code{NULL}, the password will be taken from \link{login_CopHub} inputs or, if \code{login_CopHub} is not in use, asked interactively. Not needed for connections to the pre-operational hub.}

\item{verbose}{logical, if \code{TRUE}, details on the function's progress will be visibile on the console. Default is TRUE.}
}
\value{
A data frame; each row represents one dataset. The data frame can be further filtered by its columnwise attributes. The selected rows can be handed over to the other getLandsat functions for previewing or downloading.
}
\description{
\code{getLandsat_query} queries the USGS Earth Explorer for Landsat data by some basic input search parameters. The function returns a data frame that can be further filtered.
}
\seealso{
\link{getLandsat_names} \link{getLandsat_preview} \link{getLandsat_data}
}
\author{
Jakob Schwalb-Willmann
}
