% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMODIS_data.R
\name{getMODIS_data}
\alias{getMODIS_data}
\title{Download MODIS data}
\usage{
getMODIS_data(records, dir_out = NULL, force = FALSE, n.retry = 3,
  verbose = TRUE)
}
\arguments{
\item{records}{data.frame, one or multiple records (each represented by one row), as it is returned by \link{getMODIS_query}.}

\item{dir_out}{character, full path to download target directory. Optional. If not set, \code{getMODIS_data} uses the directory to the \code{getSpatialData} archive folder. Use \link{set_archive} to once define a getSpatialData archive folder.}

\item{force}{logical. If \code{TRUE}, download is forced even if file already exisits in the download directory. Default is \code{FALSE}.}

\item{n.retry}{numeric, maximum number of download (re-)attempts. If downloads of datasets fail (e.g. MD5 checksums do not match), these downloads will be reattampted.}

\item{verbose}{logical, whether to display details on the function's progress or output on the console.}
}
\value{
Character vector of paths to the downloaded files.
}
\description{
\code{getMODIS_data} downloads MODIS data queried using \link{getMODIS_query} from different sources.
}
\details{
\code{getMODIS_data} downloads MODIS data from the Level-1 and Atmosphere Archive & Distribution System (LAADS) of NASA's Distributed Active Archive Center (DAAC) at the Goddard Space Flight Center in Greenbelt, Maryland (\url{https://ladsweb.modaps.eosdis.nasa.gov/}).
}
\examples{
## Load packages
library(getSpatialData)
library(sf)

## set aoi and time range for the query
set_aoi(aoi_data[[1]])
time_range <-  c("2017-08-01", "2017-08-30")

## Login to USGS ERS
\dontrun{
login_USGS("username")

## set archive directory
set_archive("/path/to/archive/")

## get available products and select one
product_names <- getMODIS_names()
product <- grep("MOD13Q1", product_names, value = T)

## query for records for your AOI, time range and product
query <- getMODIS_query(time_range = time_range, name = product)

## preview a record
getMODIS_preview(query[1,])

## download records 1 and 2
files <- getMODIS_data(query[1:2,])
}

}
\seealso{
\link{getMODIS_names} \link{getMODIS_query} \link{getMODIS_preview}
}
\author{
Jakob Schwalb-Willmann
}
