% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMODIS_names.R
\name{getMODIS_names}
\alias{getMODIS_names}
\title{Get MODIS dataset names from USGS Earth Explorer}
\usage{
getMODIS_names(username = NULL, password = NULL)
}
\arguments{
\item{username}{character, a valid USGS user name. Default is NULL. Leave it undefined, if you want to use \link{login_USGS} to define the login credentials once for all \code{get*} functions that connect to USGS services during the session. Register on \url{https://ers.cr.usgs.gov/register/}.}

\item{password}{character, the password to the specified user account. If \code{NULL}, the password will be taken from \link{login_USGS} inputs or, if \code{login_USGS} is not in use, asked interactively.}
}
\value{
A character vector
}
\description{
\code{getMODIS_names} obtains names of available MODIS datasets from USGS Earth Explorer. They can optionally be used with the \link{getMODIS_query} function for querying a specific MODIS dataset instead of all.
}
\seealso{
\link{getMODIS_query}
}
\author{
Jakob Schwalb-Willmann
}
