% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSentinel_data.R
\name{getSentinel_data}
\alias{getSentinel_data}
\title{Download Sentinel-1, -2 and -3 datasets by UUIDs}
\usage{
getSentinel_data(products, dir_out, hub_user, hub_pass = NULL,
  hub_access = "auto")
}
\arguments{
\item{products}{data.frame, one or multiple prodcuts (each represented by one row), as it is returned by \link{getSentinel_query}.}

\item{dir_out}{character, full path to download target directory}

\item{hub_user}{character, a valid user name to the Copernicus Open Access Hub. Register on \url{https://scihub.copernicus.eu/}. Not needed for connections to the pre-operational hub.}

\item{hub_pass}{character, the password to the specified user account. If \code{NULL}, the password will be asked interactively (default). Not needed for connections to the pre-operational hub.}

\item{hub_access}{character, either "auto" to access the Copernicus Open Access Hubs by \code{platform} input, "operational" to look for ESA's operational products from the Open Hub,  "pre-ops" to look for pre-operational products from the Pre-Ops Hub (e.g. currently all Sentinel-3 products), or an valid API URL. Default is "auto".}
}
\value{
List of files that had been downloaded
}
\description{
\code{getSentinel_data} downloads complete Sentinel datasets including imagery and meta files from the Copernicus Open Access Hubs for Sentinel data. The datasets are identified by their UUIDs, which can be extracted using \link{getSentinel_query}.
}
\details{
The \code{getSentinel*} function bundle makes use of the python library \code{sentinelsat}, serving as interface to the SciHub API. Python needs to be installed on your system.
}
\examples{
## Load packages
library(getSpatialData)
library(raster)

## Define an extent, a time range and a platform
ext <- extent(10.29048,11.75558,45.93350,46.94617)
time_range <-  c("20170801", "20170803")
platform <- "Sentinel-2"

## Define your Hub credentials
hub_user <- "your_username"

## Use getSentinel_query to search for data
\dontrun{
products <- getSentinel_query(ext = ext, time_range = time_range, platform = platform,
                              hub_user = hub_user)
#if you do not want to retype your password for every call, use the 'hub_pass' argument

## Get an overview of the products
View(products) #get an overview about the search products
colnames(products) #see all available filter attributes
unique(products$processinglevel) #use one of the, e.g. to see available processing levels

## Preview a single product
getSentinel_preview(product = products[1,], hub_user = hub_user)

## Filter the products
products_filtered <- products[which(products$processinglevel == "Level-1C"),] #filter by Level

## Download datasets
#dir_out <- "your/output/directory"
dir_out <- tempdir() #or some temporary directory for this example
files <- getSentinel_data(products = products_filtered, dir_out = dir_out,
                          hub_user = hub_user)
}
}
\seealso{
\link{getSentinel_query}
}
\author{
Jakob Schwalb-Willmann
}
