% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSentinel_data.R
\name{getSentinel_data}
\alias{getSentinel_data}
\title{Download Sentinel-1, -2 and -3 datasets by UUIDs}
\usage{
getSentinel_data(uuid, dir_out, hub_user, hub_pass = NULL,
  hub_access = "operational", py_path = NULL)
}
\arguments{
\item{uuid}{character vector, one or multiple UUIDs identifying the datasets to be downloaded.}

\item{dir_out}{character, full path to download target directory}

\item{hub_user}{character, a valid user name to the Copernicus Open Access Hub. Register on \url{https://scihub.copernicus.eu/}.}

\item{hub_pass}{character, the password to the specified user account. If \code{NULL}, the password will be asked interactively (default).}

\item{hub_access}{character, either "operational" to look for ESA's operational products from the Open Hub,  "pre-ops" to look for pre-operational products from the Pre-Ops Hub (currently all Sentinel-3 products), or an URL to a Hub. Default is "operational".}

\item{py_path}{character, define path to the python command manually. If \code{NULL}, python is searched for automatically (default). To be used, of multiple versions of Python are installed (e.g. Anaconda).}
}
\value{
List of files that had been downloaded
}
\description{
\code{getSentinel_data} downloads complete Sentinel datasets including imagery and meta files from the Copernicus Open Access Hubs for Sentinel data. The datasets are identified by their UUIDs, which can be extracted using \link{getSentinel_query}.
}
\details{
The \code{getSentinel*} function bundle makes use of the python library \code{sentinelsat}, serving as interface to the SciHub API. Python needs to be installed on your system.
}
\examples{

## Load packages
library(getSpatialData)
library(raster)

## Define an extent, a time range and a platform
ext <- extent(10.29048,11.75558,45.93350,46.94617)
time_range <-  c("20170801", "20170803")
platform <- "Sentinel-2"

## Define your Hub credentials
hub_user <- "your_account"

## Use getSentinel_query to search for data
\dontrun{
results <- getSentinel_query(ext = ext, time_range = time_range, platform = platform,
                             hub_user = hub_user)

## View the results
View(results) #get an overview about the search results
colnames(results) #see all available filter attributes
unique(results$processinglevel) #use one of the, e.g. to see available processing levels

## Filter the results
results_filtered <- results[which(results$processinglevel == "Level-1C"),] #filter by Level

## Extract UUIDs
uuid <- results_filtered$uuid #get UUIDs of your filtered query results
uuid <- uuid[c(1:3)] #take the first three datasets to be downloaded

## Download datasets
#dir_out <- "your/output/directory"
dir_out <- tempdir() #or some temporary directory for this example
files <- getSentinel_data(uuid = uuid, dir_out = dir_out,
                          hub_user = hub_user)
}
}
\seealso{
\link{getSentinel_query}
}
\author{
Jakob Schwalb-Willmann
}
