% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSentinel_restore.R
\name{getSentinel_restore}
\alias{getSentinel_restore}
\title{Restore Sentinel datasets from Copernicus LTA}
\usage{
getSentinel_restore(record, username = NULL, password = NULL,
  hub = "auto", verbose = TRUE)
}
\arguments{
\item{record}{data.frame, single row data.frame collected from the return of \link{getSentinel_query}, representing the selected record and all its attributes.}

\item{username}{character, a valid user name to the ESA Copernicus Open Access Hub. If \code{NULL} (default), the session-wide login credentials are used (see \link{login_CopHub} for details on registration).}

\item{password}{character, the password to the specified user account. If \code{NULL} (default) and no seesion-wide password is defined, it is asked interactively ((see \link{login_CopHub} for details on registration).}

\item{hub}{character, either "auto" to access the Copernicus Open Access Hubs by \code{platform} input, "operational" to look for ESA's operational products from the Open Hub,  "pre-ops" to look for pre-operational products from the Pre-Ops Hub (e.g. currently all Sentinel-3 products), or an valid API URL. Default is "auto".}

\item{verbose}{logical, if \code{TRUE}, details on the function's progress will be visibile on the console. Default is TRUE.}
}
\value{
TRUE on success.
}
\description{
\code{getSentinel_restore} requests to restore Setninel datasets that have been archived by ESA to the Copernicus Long-Term Archive (LTA) (see argument \code{check_avail} of \link{getSentinel_query}).
}
\seealso{
\link{getSentinel_query}
}
\author{
Jakob Schwalb-Willmann
}
