% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gSD_settings.R
\name{gSD_settings}
\alias{gSD_settings}
\alias{set_python}
\alias{set_cophub_login}
\title{getSpatialData settings}
\usage{
set_python(py_path)

set_cophub_login(hub_user, hub_pass = NULL)
}
\arguments{
\item{py_path}{character, path to preferred python installation.}

\item{hub_user}{character, user name to the Copernicus Open Access Hub. Register on \url{https://scihub.copernicus.eu/}.}

\item{hub_pass}{character, password to the Copernicus Open Access Hub.}
}
\value{
None.
}
\description{
Defines global settings that are used by all \code{getSpatialData} functions
}
\details{
\code{set_python} sets up the connection to a python installation manually. This prevents getSpatialData from using a wrong Pyhton version, if multiple installs are available or if the correct install is not recognized automatically.

\code{set_cophub_login} defines the login credentials for the Copernicus Open Access Hub. These will be saved and made available for all \code{getSpatialData} functions during the whole session and will be erased when quitting the session. Leave the \code{hub_user} and \code{hub_pass} arguments of the \code{getSentinel*} functions to their default value (\code{NULL}) to force them to use the global credentials defined with \code{set_cophub_login}.
}
\examples{
## Define a python path
set_python("path/to/my/python") #needs to lead to an executable file, on windows, a python.exe

## Define user credentials for the Copernicus Open Access Hub
set_cophub_login(hub_user = "my_user_name", hub_pass = "my_password")

}
\seealso{
getSentinel_query
}
\author{
Jakob Schwalb-Willmann
}
