% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gSD_settings.R
\name{gSD_settings}
\alias{gSD_settings}
\alias{set_login_CopHub}
\alias{set_archive}
\title{getSpatialData settings}
\usage{
set_login_CopHub(hub_user, hub_pass = NULL)

set_archive(dir_archive)
}
\arguments{
\item{hub_user}{character, user name to the Copernicus Open Access Hub. Register on \url{https://scihub.copernicus.eu/}.}

\item{hub_pass}{character, password to the Copernicus Open Access Hub.}

\item{dir_archive}{character, directory to the \code{getSpatialData} archive folder.}
}
\value{
None.
}
\description{
Defines global settings that are used by all \code{getSpatialData} functions
}
\details{
\code{set_login_CopHub} defines the login credentials for the Copernicus Open Access Hub. These will be saved and made available for all \code{getSpatialData} functions during the whole session and will be erased when quitting the session. Leave the \code{hub_user} and \code{hub_pass} arguments of the \code{getSentinel*} functions to their default value (\code{NULL}) to force them to use the global credentials defined with \code{set_login_CopHub}.
\code{set_archive} globally defines the directory on your machine (or an external device) where getSpatialData should build up its donwload data archive. Since getSpatialData handles big amounts of data, it is recommended to once define a location where enough free storage is available and then afterwards to not change the archive location. You need to define the archives location for each session after loading getSpatialData. It will then be remembered for the duration of the session. Apart from the archive location, you can manually define a download path when calling the *_data functions. If you do not define a path there, getSpatialData will direct the download to the defined archive. The archive is structred by sensors.
}
\examples{
## Define user credentials for the Copernicus Open Access Hub
set_login_CopHub(hub_user = "my_user_name", hub_pass = "my_password")

}
\seealso{
getSentinel_query
}
\author{
Jakob Schwalb-Willmann
}
