% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_product_checker.R
\name{is.landsat}
\alias{is.landsat}
\title{Returns TRUE for records that are of product group 'Landsat'}
\usage{
is.landsat(records)
}
\arguments{
\item{records}{data.frame, one or multiple records (each represented by one row), as it is 
returned by \link{get_records}.}
}
\value{
logical vector, same length as number of rows in \code{records}.
}
\description{
\code{is.landsat} checks which records are Landsat records.
}
\author{
Henrik Fisser, 2020
}
