% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_product_checker.R
\name{is.landsat8}
\alias{is.landsat8}
\title{Returns TRUE for records that are of product 'LANDSAT_8_C1' (Landsat-8)}
\usage{
is.landsat8(records)
}
\arguments{
\item{records}{data.frame, one or multiple records (each represented by one row), as it is 
returned by \link{get_records}.}
}
\value{
logical vector, same length as number of rows in \code{records}.
}
\description{
\code{is.landsat8} checks which records are Landsat-8 records.
}
\author{
Henrik Fisser, 2020
}
