% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_product_checker.R
\name{is.product_group_}
\alias{is.product_group_}
\title{---------------------------------------------------------------------
name: is_product_checker
description: These are check utils for product and product group and
work in same manner as is.na(). Apart from product and product group
some methods target more specific types that are identified from 
the record ids. All methods return logical vectors of the same length as NROW(records).
author: Henrik Fisser, 2020
---------------------------------------------------------------------
Returns TRUE for records that are of the specified product group}
\usage{
is.product_group_(records, product_group)
}
\arguments{
\item{records}{sf data.frame}

\item{product_group}{character specifies the product group to be checked on.}
}
\value{
logical vector
}
\description{
\code{is.product_group_} checks which records are of product group \code{product_group}
}
\author{
Henrik Fisser, 2020
}
