% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_previews.R
\name{view_previews}
\alias{view_previews}
\alias{plot_previews}
\title{View or plot previews of records}
\usage{
view_previews(records, show_aoi = TRUE, aoi_colour = "red",
  value = FALSE, verbose = TRUE)

plot_previews(records, show_aoi = TRUE, aoi_colour = "red",
  separate = FALSE, maxcol = 3, value = FALSE, verbose = TRUE)
}
\arguments{
\item{records}{records data frame, containing one or multiple records (each represented by one row), as returned by \link{get_records}}

\item{show_aoi}{logical, whether the session AOI defined with \link{set_aoi} should be drawn on the map or plot(s) (\code{TRUE}) or not.}

\item{aoi_colour}{character, colour of the AOI. Ignored, if \code{show_aoi = FALSE}.}

\item{value}{logical, whether to return the displayed map/plot object (\code{TRUE}) or not.}

\item{verbose}{logical, whether to display details on the function's progress or output on the console.}

\item{separate}{logical, whether previews should be displayed separately on a \code{cowplot} grid (\code{TRUE}) or as a mosaic in a single plot (\code{FALSE}, default).}

\item{maxcol}{numeric, maximum number of columns that the plot grid should have.}
}
\value{
None or a \code{mapview}/\code{plot} object
}
\description{
\code{view_previews} and \code{plot_previews} load and display georeferenced previews acquired using \code{\link{get_previews}}.
}
\details{
\code{view_previews} renders previews as RGB images on an interactive map using \code{mapview} and \code{leaflet}.

\code{plot_previews} plots previews to a graphics device using \code{ggplot2}.
}
\author{
Jakob Schwalb-Willmann
}
