% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_records.R
\name{view_records}
\alias{view_records}
\alias{plot_records}
\title{View footprints of records}
\usage{
view_records(
  records,
  show_aoi = TRUE,
  line_colours = hcl.colors(nrow(records)),
  fill_colours = hcl.colors(nrow(records)),
  fill_alpha = 0.3,
  aoi_colour = "deepskyblue",
  verbose = TRUE
)

plot_records(
  records,
  show_aoi = TRUE,
  line_colours = hcl.colors(nrow(records)),
  fill_colours = hcl.colors(nrow(records)),
  fill_alpha = 0.3,
  aoi_colour = "deepskyblue",
  value = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{records}{records data frame, containing one or multiple records (each represented by one row), as returned by \link{get_records}}

\item{show_aoi}{logical, whether the session AOI defined with \link{set_aoi} should be drawn on the map or plot(s) (\code{TRUE}) or not.}

\item{line_colours}{character, colours for each footprint's outline. Vector must be of same length as number of records (\code{nrow(records)}).}

\item{fill_colours}{character, colours of each footprint's fill. Vector must be of same length as number of records (\code{nrow(records)})}

\item{fill_alpha}{numeric, alpha value for \code{fill_colours}. Value must be between 0 (transparent) and 1 (opaque).}

\item{aoi_colour}{character, colour of the AOI. Ignored, if \code{show_aoi = FALSE}.}

\item{verbose}{logical, whether to display details on the function's progress or output on the console.}
}
\value{
None or a \code{mapview}/\code{plot} object
}
\description{
\code{view_records} and \code{plot_records} display the footsprint of each record on a map.
}
\author{
Jakob Schwalb-Willmann
}
