% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_records.R
\name{write_records}
\alias{write_records}
\title{Writes records to a file}
\usage{
write_records(records, file_name = NULL, driver = "GPKG",
  dir_out = NULL, file = NULL, append = FALSE, verbose = TRUE)
}
\arguments{
\item{file_name}{character name of the file to be saved. If it is NULL a file name will be 
automatically generated from date and time.}

\item{driver}{character specifies the writer. Options are all returned by
sf::st_drivers() and "csv". Will be ignored in case a fully qualified
\code{file} path is provided. Default is "GPKG"}

\item{file}{character complete file path where to write the records. Must be a full file path
including file basename and extension matching a supported driver. Default is NULL.}

\item{append}{logical, if TRUE and a file path is given through \code{file}, the record(s)
will be appended to \code{file} if possible.}

\item{verbose}{logical, whether to display details on the function's progress or output on the console.}
}
\value{
character file path to the saved records.
}
\description{
\code{write_records} write records to a file.
}
\author{
Henrik Fisser, 2020
}
